requirements.txt
Pillow
faster-whisper
ctranslate2
openpyxl
python-docx
pdfplumber
pypdf
PyPDF2
